Profiles
An interpolation of the bedrock, as well as of both englacial features has been conducted. The profiles show where the picked and the interpolated bedrock are located. In addition, the profiles show where englacial features have been picked and interpolated, respectively. Table 3 in the Appendix shows where the bedrock, and englacial Horizons have been picked.